# Improve performance of hashRegExp lookup