@lvivski 2023-03-10
@ryanwilsonperkin 

> Could you include some information about the format of the hashes

we're using `xxxhash64` 8 chars.

> the number and size of assets that you're replacing them in

~5 runtime assets
~10k hashes to replace. 5k from content hash and 5k through the hook with [`webpack-subresource-integrity`](https://github.com/waysact/webpack-subresource-integrity). I've also tried disabling SRI and it didn't help.

> And are you able to isolate the timeout to this part of the code?

Yes, after creating a custom `webpack` build with this change reverted, it works fine.


I have just ran your stress-test-build from https://github.com/ryanwilsonperkin/webpack-16758

```
CPU: 2.4 GHz 8-Core Intel Core i9
RAM: 64 GB 2667 MHz DDR4
OS: Mac OS 13.2.1
NodeJS: v19.7.0
```

### `v5.76.0`

```
Run 1: 50487 ms asset processing > RealContentHashPlugin
Run 2: 51002 ms asset processing > RealContentHashPlugin
Run 3: 51650 ms asset processing > RealContentHashPlugin
```

### `v5.76.0 w/ changes from this PR reverted`

```
Run 1: 9812 ms asset processing > RealContentHashPlugin
Run 2: 10482 ms asset processing > RealContentHashPlugin
Run 3: 10761 ms asset processing > RealContentHashPlugin
```

Around 5x decrease in processing time if I revert this change.