@ryanwilsonperkin 2023-03-10
I'm getting very different results than you on that stress test build:

```
CPU: Apple M1 Pro
RAM: 16 GB
OS: Mac OS 13.2.1
NodeJS: v19.4.0
```

## `v5.75.0`
```
Run 1: 46028 ms asset processing > RealContentHashPlugin
Run 2: 46286 ms asset processing > RealContentHashPlugin
Run 3: 45108 ms asset processing > RealContentHashPlugin
```

## `v5.76.0`
```
Run 1: 25504 ms asset processing > RealContentHashPlugin
Run 2: 26187 ms asset processing > RealContentHashPlugin
Run 3: 27294 ms asset processing > RealContentHashPlugin
```

Your machine appears to be more powerful, but it seems odd to me that when running the test from https://github.com/ryanwilsonperkin/webpack-16758 that your results would be twice as slow against `5.76.0` but four times as fast against `5.75.0`. 

> the number and size of assets that you're replacing them in

I should clarify this further, I meant the size (in bytes) of the assets that you are replacing the hashes in, and how many of those assets you're processing.