@lvivski 2023-03-10
Sorry, I should have clarified that I wasn't running `5.7_6_` against `5.7_5_` on your test setup . I was comparing `5.7_6_` and then manually remove changes from this PR, but keeping all other changes from `5.7_6_`.


I wonder if this is `M1` vs `Intel` difference. Do you use any CI? If so, do you also see similar improvement numbers there?

> I meant the size (in bytes) of the assets that you are replacing the hashes in

This is generally hard to tell, ranging between several Kb to low hundreds, with about 150Mb generated JS total. 

But yeah, I think that running multiple small RexExp's is less efficient than a single big one, in our case, because of the total number and size of the files. While initialization time might be longer for a big regex, subsequents comparisons are faster with only a single pass through the each file.