@lvivski 2023-03-10
I've created a setup with random strings in files that can repro my issue. https://gist.github.com/lvivski/b3245a494be6a862b63dc4e00012eb1b

### `v5.7*5*.0`
```
Run 1: 79448 ms asset processing > RealContentHashPlugin
Run 2: 80579 ms asset processing > RealContentHashPlugin
```

### `v5.7*6*.0`
```
Run 1: 399699 ms asset processing > RealContentHashPlugin
Run 2: 418393 ms asset processing > RealContentHashPlugin
```

### `v5.7*6*.0` w/ this PR reverted
```
Run 1: 77584 ms asset processing > RealContentHashPlugin
Run 2: 78269 ms asset processing > RealContentHashPlugin
```

So it seems that the root cause is this change.