@vankop 2023-03-10
@alexander-akait @TheLarkInn could we test this one more time.. problem here is that if you have small amount of chunks (not a 3000 as in @ryanwilsonperkin example 😀 ) => it is fine to concat them + matching on huge chunk content would be better. So what this changes does in reality, if you have 90kb chunk and 20 refs ( hashes ) => webpack creates 20 regexps and run then 20 times on 90kb chunk.